Source value can be nil to remove rendered component.
Previously only the initial value of a live component's source could
be nil and a placeholder will be rendered instead of the component.

Now the source value can change back to nil and the component will be
replaced with a placeholder.

This is because we aren't using go blocks as the conveyence mechanism anymore.