Add unlisten cleanup for live collection
As live-collection is not registered with a source, but the source
listener is added separately, it is not subject to automatic unlisten
when cleaning up. We need to add a separate cleanup that removes
source listener.

This is only a problem with long-lived sources (like an atom) that
multiple pages might be listening to.